new types file for Modal, refactor the code
